fix bugs and add constraint for application
