+++ app/linux/runner/my_application.cc
@@ -7,6 +7,9 @@
 
 #include "flutter/generated_plugin_registrant.h"
 
+// Library for customizing the application window.
+#include <bitsdojo_window_linux/bitsdojo_window_plugin.h>
+
 struct _MyApplication {
   GtkApplication parent_instance;
   char** dart_entrypoint_arguments;
@@ -47,7 +50,9 @@ static void my_application_activate(GApplication* application) {
     gtk_window_set_title(window, "gameverse");
   }
 
-  gtk_window_set_default_size(window, 1280, 720);
+  auto bdw = bitsdojo_window_from(window);            // <--- add this line
+  bdw->setCustomFrame(true);                          // <-- add this line
+  //gtk_window_set_default_size(window, 1280, 720);   // <-- comment this line
   gtk_widget_show(GTK_WIDGET(window));
 
   g_autoptr(FlDartProject) project = fl_dart_project_new();
