+++ app/macos/Runner/MainFlutterWindow.swift
@@ -1,7 +1,11 @@
 import Cocoa
 import FlutterMacOS
+import bitsdojo_window_macos // Add this line
 
-class MainFlutterWindow: NSWindow {
+class MainFlutterWindow: BitsdojoWindow {
+  override func bitsdojo_window_configure() -> UInt {
+    return BDW_CUSTOM_FRAME | BDW_HIDE_ON_STARTUP
+  }
   override func awakeFromNib() {
     let flutterViewController = FlutterViewController()
     let windowFrame = self.frame
