+++ app/pubspec.yaml
@@ -55,6 +55,9 @@ dependencies:
   # Routing
   go_router: ^15.2.0
 
+  # Customize size application
+  bitsdojo_window: ^0.1.6
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
